full patter and viz is working

+++ Arduino.h
@@ -1,5 +1,5 @@
 #pragma once
 
-long int random(int a, int b);
+long int random(uint16_t a, uint16_t b);
 void randomSeed(unsigned long seed);
 void delayMicroseconds(unsigned long us);
